task ref [9277,9278,9279,9280,9281,9284,9285,9298,9299,9300,9301,9303] #21

+++ src/main/java/com/ra/model/dto/res/ProductDetailResponse.java
@@ -16,7 +16,7 @@ public class ProductDetailResponse {
     private String image;
     private String productDetailName;
     private boolean status;
-    private int stock;
+    private Long stock;
     private double unitPrice;
     private String color;
     private Long productId;
